+++ src/main/java/net/litetex/capes/provider/MinecraftCapesCapeProvider.java
@@ -1,12 +1,13 @@
 package net.litetex.capes.provider;
 
 import java.io.IOException;
+import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.util.Map;
 
 import com.google.gson.Gson;
+import com.google.gson.annotations.SerializedName;
 import com.mojang.authlib.GameProfile;
 
 import net.litetex.capes.handler.textures.AnimatedSpriteTextureResolver;
@@ -43,33 +44,53 @@ public class MinecraftCapesCapeProvider implements CapeProvider
 		final GameProfile profile) throws IOException, InterruptedException
 	{
 		requestBuilder
-			.setHeader("User-Agent", "minecraftcapes-mod/" + SharedConstants.getCurrentVersion().name())
-			.setHeader("Accept", "application/json");
+			.setHeader("User-Agent", "minecraftcapes-mod/" + SharedConstants.getCurrentVersion().name());
 		
+		record ResponseData(
+			Boolean animatedCape,
+			@SerializedName("animated_cape_url")
+			String animatedCapeUrl,
+			@SerializedName("cape_url")
+			String capeUrl
+		)
+		{
+		}
+		
+		final ResponseData responseData;
 		try(final HttpClient client = clientBuilder.build())
 		{
 			final HttpResponse<String> response =
-				client.send(requestBuilder.GET().build(), HttpResponse.BodyHandlers.ofString());
+				client.send(
+					requestBuilder.copy()
+						.setHeader("Accept", "application/json")
+						.GET()
+						.build(),
+					HttpResponse.BodyHandlers.ofString());
 			
 			if(response.statusCode() / 100 != 2)
 			{
 				return null;
 			}
 			
-			record ResponseData(
-				Boolean animatedCape,
-				Map<String, String> textures
-			)
-			{
-			}
-			
-			final ResponseData responseData = new Gson().fromJson(response.body(), ResponseData.class);
-			
-			return new ResolvedTextureInfo.Base64TextureInfo(
-				responseData.textures().get("cape"),
-				responseData.animatedCape() ? AnimatedSpriteTextureResolver.ID : null
-			);
+			responseData = new Gson().fromJson(response.body(), ResponseData.class);
 		}
+		if(responseData == null)
+		{
+			return null;
+		}
+		
+		final String textureUrl = Boolean.TRUE.equals(responseData.animatedCape())
+			? responseData.animatedCapeUrl()
+			: responseData.capeUrl();
+		if(textureUrl == null)
+		{
+			return null;
+		}
+		
+		return CapeProvider.resolveTextureDefault(
+			clientBuilder,
+			requestBuilder.copy().uri(URI.create(textureUrl)),
+			responseData.animatedCape() ? AnimatedSpriteTextureResolver.ID : null);
 	}
 	
 	@Override