+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.2.1
+* Fix MinecraftCapes Provider no longer working due to breaking API changes #206
+* Removed Wynntils #204
+
 # 4.2.0
 * Updated to 1.21.11